move metrics.rb to bundler/lib as it no longer interacts with fetcher
